refactor of compute shader

+++ src/util.rs
@@ -1,4 +1,4 @@
-use core::f64;
+use core::f32;
 use std::{
     io::{self, Write},
     ops::RangeInclusive,
@@ -9,7 +9,7 @@ use rand::{Rng, rng};
 use crate::math::Vec3;
 
 pub fn print_progress(j: i32, height: i32) {
-    let progress = (100.0 * (j as f64 + 1.0) / height as f64) as i32;
+    let progress = (100.0 * (j as f32 + 1.0) / height as f32) as i32;
     assert!(progress >= 0 && progress <= 100);
     let plus = "#".repeat(progress as usize / 4);
     let minus = "-".repeat(25 - (progress as usize / 4));
@@ -17,12 +17,12 @@ pub fn print_progress(j: i32, height: i32) {
     io::stdout().flush().unwrap();
 }
 
-pub fn random_float() -> f64 {
+pub fn random_float() -> f32 {
     let mut rng = rng();
     rng.random_range(0.0..=1.0)
 }
 
-pub fn random_float_range(min: f64, max: f64) -> f64 {
+pub fn random_float_range(min: f32, max: f32) -> f32 {
     let mut rng = rng();
     rng.random_range(min..=max)
 }
@@ -31,13 +31,13 @@ pub fn sample_square() -> Vec3 {
     Vec3::new(random_float() - 0.5, random_float() - 0.5, 0.0)
 }
 
-pub fn linear_to_gamma(linear: f64) -> f64 {
+pub fn linear_to_gamma(linear: f32) -> f32 {
     if linear > 0.0 {
-        return f64::sqrt(linear);
+        return f32::sqrt(linear);
     }
     0.0
 }
 
-pub fn surrounds(range: &RangeInclusive<f64>, a: &f64) -> bool {
+pub fn surrounds(range: &RangeInclusive<f32>, a: &f32) -> bool {
     range.start() < a && range.end() > a
 }
